Move httputils/mimtype to the single consumser, and remove unused function.
